Grind charaobj: signed frame-counter modulo in StaticFrame
System.m_frameCounter % 30 is computed with signed division in the target
(mulhw + srawi). Cast the counter to int so the modulo matches the
target's signed magic-number division instead of the unsigned mulhwu form.

StaticFrame 81.75% -> 88.25% (+5 matched instrs). Unit 81.59% -> 81.64%.

+++ src/charaobj.cpp
@@ -3278,7 +3278,7 @@ void CGCharaObj::StaticFrame()
 			unsigned short hp = *reinterpret_cast<unsigned short*>(script + 0x1C);
 			unsigned short maxHp = *reinterpret_cast<unsigned short*>(script + 0x1A);
 			if (hp != 0 && hp <= (maxHp >> 2)) {
-				if ((System.m_frameCounter % 0x1E) == 0) {
+				if ((static_cast<int>(System.m_frameCounter) % 0x1E) == 0) {
 					Sound.PlaySe(0x53, 0x40, 0x7F, 0);
 				}
 				break;
